[fix]: 관심사 분리 정리
