+++ core/network/src/main/java/com/hmh/hamyeonham/core/network/di/AuthModule.kt
@@ -1,7 +1,7 @@
 package com.hmh.hamyeonham.core.network.di
 
 import com.hmh.hamyeonham.common.qualifier.Unsecured
-import com.hmh.hamyeonham.core.network.login.LoginService
+import com.hmh.hamyeonham.core.network.login.AuthService
 import dagger.Module
 import dagger.Provides
 import dagger.hilt.InstallIn
@@ -15,5 +15,5 @@ import javax.inject.Singleton
 object AuthModule {
     @Provides
     @Singleton
-    fun provideLoginApi(@Unsecured retrofit: Retrofit): LoginService = retrofit.create()
+    fun provideLoginApi(@Unsecured retrofit: Retrofit): AuthService = retrofit.create()
 }
