--- data/login/src/main/java/com/hmh/hamyeonham/login/DefaultLogoutRepository.kt
@@ -1,16 +1,0 @@
-package com.hmh.hamyeonham.login
-
-import com.hmh.hamyeonham.core.network.login.LoginService
-import com.hmh.hamyeonham.login.repository.LogoutRepository
-import javax.inject.Inject
-
-class DefaultLogoutRepository @Inject constructor(
-    private val logoutService: LoginService,
-) : LogoutRepository {
-
-    override suspend fun logout(): Result<Unit> {
-        return runCatching {
-            logoutService.logout()
-        }
-    }
-}
