--- data/login/src/main/java/com/hmh/hamyeonham/login/DefaultSignUpRepository.kt
@@ -1,27 +1,0 @@
-package com.hmh.hamyeonham.login
-
-import com.hmh.hamyeonham.core.network.service.SignUpService
-import com.hmh.hamyeonham.core.network.signup.model.toSignUpRequest
-import com.hmh.hamyeonham.login.model.SignRequestDomain
-import com.hmh.hamyeonham.login.model.SignUpUser
-import com.hmh.hamyeonham.login.repository.SignUpRepository
-import javax.inject.Inject
-
-class DefaultSignUpRepository @Inject constructor(
-    private val signUpService: SignUpService,
-) : SignUpRepository {
-
-    override suspend fun signUp(
-        accessToken: String,
-        signUpRequest: SignRequestDomain,
-    ): Result<SignUpUser> {
-        return runCatching {
-            val bearerToken = "Bearer $accessToken"
-            signUpService.signUp(
-                bearerToken,
-                "Android",
-                signUpRequest.toSignUpRequest()
-            ).data.toSignUpUser()
-        }
-    }
-}
