+++ data/login/src/main/java/com/hmh/hamyeonham/login/di/AuthModule.kt
@@ -1,7 +1,7 @@
 package com.hmh.hamyeonham.login.di
 
-import com.hmh.hamyeonham.login.DefaultSignUpRepository
-import com.hmh.hamyeonham.login.repository.SignUpRepository
+import com.hmh.hamyeonham.login.DefaultAuthRepository
+import com.hmh.hamyeonham.login.repository.AuthRepository
 import dagger.Binds
 import dagger.Module
 import dagger.hilt.InstallIn
@@ -10,12 +10,12 @@ import javax.inject.Singleton
 
 @Module
 @InstallIn(SingletonComponent::class)
-object SignUpModule {
+object AuthModule {
     @Module
     @InstallIn(SingletonComponent::class)
     interface Binder {
         @Binds
         @Singleton
-        fun provideSignUpRepository(signRepository: DefaultSignUpRepository): SignUpRepository
+        fun provideUsageGoalsRepository(loginRepository: DefaultAuthRepository): AuthRepository
     }
 }
