+++ domain/login/src/main/java/com/hmh/hamyeonham/login/repository/AuthRepository.kt
@@ -1,8 +1,12 @@
 package com.hmh.hamyeonham.login.repository
 
+import com.hmh.hamyeonham.login.model.Login
 import com.hmh.hamyeonham.login.model.SignRequestDomain
 import com.hmh.hamyeonham.login.model.SignUpUser
 
-interface SignUpRepository {
+interface AuthRepository {
+    suspend fun login(accessToken: String): Result<Login>
+    suspend fun logout(): Result<Unit>
     suspend fun signUp(accessToken: String, signUpRequest: SignRequestDomain): Result<SignUpUser>
+    suspend fun withdrawal(): Result<Unit>
 }
