--- domain/login/src/main/java/com/hmh/hamyeonham/login/repository/LoginRepository.kt
@@ -1,7 +1,0 @@
-package com.hmh.hamyeonham.login.repository
-
-import com.hmh.hamyeonham.login.model.Login
-
-interface LoginRepository {
-    suspend fun login(accessToken: String): Result<Login>
-}
