--- domain/login/src/main/java/com/hmh/hamyeonham/login/repository/LogoutRepository.kt
@@ -1,6 +1,0 @@
-package com.hmh.hamyeonham.login.repository
-
-import com.hmh.hamyeonham.login.model.Login
-interface LogoutRepository {
-    suspend fun logout(): Result<Unit>
-}
