--- domain/login/src/main/java/com/hmh/hamyeonham/login/repository/WithdrawalRepository.kt
@@ -1,5 +1,0 @@
-package com.hmh.hamyeonham.login.repository
-
-interface WithdrawalRepository {
-    suspend fun withdrawal(): Result<Unit>
-}
