+++ feature/login/src/main/java/com/hmh/hamyeonham/feature/login/LoginViewModel.kt
@@ -4,7 +4,7 @@ import android.content.Context
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import com.hmh.hamyeonham.core.network.auth.datastore.HMHNetworkPreference
-import com.hmh.hamyeonham.login.repository.LoginRepository
+import com.hmh.hamyeonham.login.repository.AuthRepository
 import com.kakao.sdk.common.model.ClientError
 import com.kakao.sdk.common.model.ClientErrorCause
 import com.kakao.sdk.user.UserApiClient
@@ -29,7 +29,7 @@ data class LoginState(
 
 @HiltViewModel
 class LoginViewModel @Inject constructor(
-    private val loginRepository: LoginRepository,
+    private val authRepository: AuthRepository,
     private val hmhNetworkPreference: HMHNetworkPreference,
 ) : ViewModel() {
 
@@ -56,7 +56,7 @@ class LoginViewModel @Inject constructor(
                     loginWithKakaoAccount(context)
                 } else if (token != null) {
                     viewModelScope.launch {
-                        loginRepository.login(token.accessToken).onSuccess {
+                        authRepository.login(token.accessToken).onSuccess {
                             _kakaoLoginEvent.emit(LoginEffect.LoginSuccess)
                             hmhNetworkPreference.accessToken = token.accessToken
                             hmhNetworkPreference.refreshToken = token.refreshToken
@@ -83,7 +83,7 @@ class LoginViewModel @Inject constructor(
                 // 닉네임 정보 얻기 실패 시
             } else if (token != null) {
                 viewModelScope.launch {
-                    loginRepository.login(token.accessToken).onSuccess {
+                    authRepository.login(token.accessToken).onSuccess {
                         _kakaoLoginEvent.emit(LoginEffect.LoginSuccess)
                         hmhNetworkPreference.run {
                             accessToken = token.accessToken
